fixed Piglin Barter probabilities
The wiki claims that the barter rates for ender pearls are increased and barters are only guaranteed for the first 72 trades. My testing shows, that ender pearl rates are slightly decreased (~4.14%) from vanilla 1.16.1 rates (~4.73%), and RedLime confirmed that the guranteed barters occur for every set of 72 trades, not only the first 72 trades.